# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/api/v1/user/login" : {
      "post" : {
        "summary" : "用户登录",
        "description" : "用户登录",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "用户名",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "密码",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "userId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "totalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "normalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "expiredCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "version" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "username" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "password" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "phone" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "email" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/register" : {
      "post" : {
        "summary" : "用户注册",
        "description" : "用户注册",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "userId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "totalCount" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "seckillCount" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "normalCount" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "expiredCount" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "version" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "id" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "username" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "password" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "phone" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "email" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "userId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "totalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "normalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "expiredCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "version" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "username" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "password" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "phone" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "email" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/profile" : {
      "get" : {
        "summary" : "获取用户信息",
        "description" : "获取用户信息",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "description" : "用户ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "integer",
                      "description" : "状态码，0 表示成功"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "userId" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "totalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "seckillCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "normalCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "expiredCount" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "updateTime" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "version" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "id" : {
                          "type" : "integer",
                          "description" : "",
                          "format" : "int64"
                        },
                        "username" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "password" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "phone" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "email" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "status" : {
                          "type" : "integer",
                          "description" : ""
                        },
                        "createTime" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : "返回数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}